+++ framework/ui/composables/index.ts
@@ -7,4 +7,5 @@ export * from "@ui/composables/useTeleportTarget";
 export * from "@ui/composables/useTableSort";
 export * from "@ui/composables/useTableSelection";
 export * from "@ui/composables/useDataTableSort";
+export * from "@ui/composables/useDataTablePagination";
 export * from "@ui/composables/useBladeLoading";